pyproject.toml
[tool.poetry]
name = "handmade"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.6.9"
Django = "^3.0.5"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.2.1"
Pillow = "^7.1.0"
django-filter = "^2.2.0"
channels = "^2.4.0"
channels_redis = "^3.1.0"
hiredis = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
